fix: поправить ответы на запросы согласно задания

+++ controllers/movies.js
@@ -1,5 +1,18 @@
 const Movie = require('../models/movie');
+
 const NotFoundError = require('../errors/not-found-error');
+const AccessError = require('../errors/access-error');
+const BadRequestError = require('../errors/bad-request-error');
+
+const hendlerError = (err) => {
+  switch (err.name) {
+    case 'ValidationError':
+      return new BadRequestError(`${Object.values(err.errors).map((error) => error.message).join(', ')}`);
+    case 'CastError':
+      return new BadRequestError(`Ошибка запроса ${err.message}`);
+    default: return err;
+  }
+};
 
 module.exports.createMovie = (req, res, next) => {
   const {
@@ -32,8 +45,13 @@ module.exports.createMovie = (req, res, next) => {
     movieId,
     owner: currentUserId,
   })
-    .then((createdMovie) => res.send(createdMovie))
-    .catch(next);
+    .then((createdMovie) => {
+      const movieWithoutOwner = createdMovie;
+      movieWithoutOwner.owner = undefined;
+
+      res.send(movieWithoutOwner);
+    })
+    .catch((err) => next(hendlerError(err)));
 };
 
 module.exports.getMoviesByOwner = (req, res, next) => {
@@ -41,19 +59,27 @@ module.exports.getMoviesByOwner = (req, res, next) => {
 
   Movie.find({ owner: currentUserId })
     .then((movies) => res.send(movies))
-    .catch(next);
+    .catch((err) => next(hendlerError(err)));
 };
 
 module.exports.removeMovie = (req, res, next) => {
   const currentUserId = req.user._id;
   const { movieId } = req.params;
 
-  Movie.find({ _id: movieId, owner: currentUserId })
-    .orFail(() => { throw new NotFoundError('Фильм по запросу не найден'); })
-    .then(() => {
+  Movie.findById({ _id: movieId }).select('+owner')
+    .orFail(() => { throw new NotFoundError('Фильм с указанным идентификатором не найден'); })
+    .then((movie) => {
+      if (!movie.owner.equals(currentUserId)) {
+        throw new AccessError('Вы можете удалять только свои карточки');
+      }
       Movie.findByIdAndDelete({ _id: movieId })
-        .then((deletedMovie) => res.send(deletedMovie))
-        .catch(next);
+        .then((deletedMovie) => {
+          const movieWithoutOwner = deletedMovie;
+          movieWithoutOwner.owner = undefined;
+
+          res.send(movieWithoutOwner);
+        })
+        .catch((err) => next(hendlerError(err)));
     })
-    .catch(next);
+    .catch((err) => next(hendlerError(err)));
 };
